+++ next-docs/pages/core/switch.tsx
@@ -3,9 +3,7 @@ import Preview from '../../components/codePreview/Preview';
 import ComponentPageDescription from '../../components/ComponentPageDescription';
 import PropsTable from '../../components/PropsTable';
 
-import SizesCode from '../../public/examples/switchTW/codeExamples/Sizes';
-import SizesWithIconsCode from '../../public/examples/switchTW/codeExamples/SizesWithIcons';
-
+import DefaultCode from '../../public/examples/switchTW/codeExamples/DefaultCode';
 import Custom from '../../public/examples/switchTW/Custom';
 import Default from '../../public/examples/switchTW/Default';
 import Disabled from '../../public/examples/switchTW/Disabled';
@@ -67,6 +65,12 @@ const Example = () => {
         title="Using with HTML forms"
         preview={<Form />}
         code={examples ? examples.Form : 'Loading'}
+      />
+
+      <Preview
+        title="Code exemples for full cistomization"
+        preview={<DefaultCode />}
+        code={examplesCode ? examplesCode.DefaultCode : 'Loading'}
       />
 
       <PropsTable
@@ -152,17 +156,6 @@ const Example = () => {
             description: 'Set icon for unchecked state',
           },
         ]}
-      />
-
-      <Preview
-        title="Code exemples for full cistomization"
-        preview={<SizesCode />}
-        code={examplesCode ? examplesCode.Sizes : 'Loading'}
-      />
-      <Preview
-        title="Code exemples with icons for full cistomization "
-        preview={<SizesWithIconsCode />}
-        code={examplesCode ? examplesCode.SizesWithIcons : 'Loading'}
       />
     </>
   );