+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Change Log
 
+## 9.7.8
+
+### Patch Changes
+
+- fix build
+- Updated dependencies [undefined]
+  - @heathmont/moon-assets@9.7.8
+  - @heathmont/moon-components@9.7.8
+  - @heathmont/moon-core@9.7.8
+  - @heathmont/moon-icons@9.7.8
+  - @heathmont/moon-themes@9.7.8
+  - @heathmont/moon-utils@9.7.8
+
+## 9.7.7
+
+### Patch Changes
+
+- udate v9.7.7
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.7.7
+  - @heathmont/moon-assets@9.7.7
+  - @heathmont/moon-components@9.7.7
+  - @heathmont/moon-icons@9.7.7
+  - @heathmont/moon-themes@9.7.7
+  - @heathmont/moon-utils@9.7.7
+
 ## 9.7.6
 
 ### Patch Changes