+++ workspaces/templates/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @heathmont/moon-templates
 
+## 9.7.8
+
+### Patch Changes
+
+- fix build
+
+## 9.7.7
+
+### Patch Changes
+
+- udate v9.7.7
+
 ## 9.7.6
 
 ### Patch Changes